86cyrzv5y: moved everything to role folder
